- up to 20090810 - rootns patch applied:   http://dev.splitbrain.org/darcsweb/darcsweb.cgi?r=dwplugins/data;a=commitdiff;h=20090810163355-7ad00-27c2dd29bddfa1d19a6285601d9202be41912b9e.gz;
Changed files:
    dokuwiki-plugin-data.spec -> 1.9
    remove-rootns.patch -> 1.2

+++ dokuwiki-plugin-data.spec
@@ -1,15 +1,14 @@
 %define		plugin		data
 Summary:	DokuWiki Structured Data Plugin
 Name:		dokuwiki-plugin-%{plugin}
-Version:	20090213
-Release:	4
+Version:	20090810
+Release:	1
 License:	GPL v2
 Group:		Applications/WWW
 Source0:	http://dev.splitbrain.org/download/snapshots/data-plugin-latest.tgz
-# Source0-md5:	6a3ee212496a60a343b62246e8002957
+# Source0-md5:	f55adb2c5643693fca8b7ef73944f6e4
 URL:		http://wiki.splitbrain.org/plugin:data
 Patch0:		interwiki.patch
-Patch1:		remove-rootns.patch
 BuildRequires:	rpmbuild(macros) >= 1.520
 Requires:	dokuwiki >= 20090214b-5
 Requires:	php(sqlite)
@@ -32,10 +31,15 @@ different to the repository plugin.
 %prep
 %setup -q -n %{plugin}
 %patch0 -p1
-%patch1 -p1
 
 # cleanup backups after patching
 find '(' -name '*~' -o -name '*.orig' ')' -print0 | xargs -0 -r -l512 rm -f
+
+version=$(awk -F"'" '/date/{print $4}' helper.php)
+if [ $(echo "$version" | tr -d -) != %{version} ]; then
+	: %%{version} mismatch
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
